pyproject.toml
[project]
name = "jayesian"
version = "0.1.0"
description = "Dark Energy Jaxoscopic Instrument"
requires-python = ">=3.13"
dependencies = [
    "anesthetic>=2.10.0",
    "blackjax",
    "tensorflow-probability>=0.25.0",
    "tqdm>=4.67.1",
    "fire",
]

[tool.uv.sources]
blackjax = { git = "https://github.com/handley-lab/blackjax", rev = "1d" }
